Refactor out comm object into the default Printer plugin.
The comm object is really part of the printer, so we will treat it as
such.  The comm helper functions will stay as part of the util, as other
implimentations of printers might want to use them as well.

AE-626
