+++ src/octoprint/plugins/octoprint_printer/comm.py
@@ -13,93 +13,20 @@ import threading
 import Queue as queue
 import logging
 import serial
-import octoprint.plugin
 
 from collections import deque
 
-from octoprint.util.avr_isp import stk500v2
-from octoprint.util.avr_isp import ispBase
+import octoprint.plugin
 
-from octoprint.settings import settings, default_settings
+from octoprint.util.avr_isp import ispBase
+from octoprint.util.avr_isp import stk500v2
+
 from octoprint.events import eventManager, Events
 from octoprint.filemanager import valid_file_type
 from octoprint.filemanager.destinations import FileDestinations
-from octoprint.util import get_exception_string, sanitize_ascii, filter_non_ascii, CountedEvent, RepeatedTimer
-
-try:
-	import _winreg
-except:
-	pass
-
-def serialList():
-	baselist=[]
-	if os.name=="nt":
-		try:
-			key=_winreg.OpenKey(_winreg.HKEY_LOCAL_MACHINE,"HARDWARE\\DEVICEMAP\\SERIALCOMM")
-			i=0
-			while(1):
-				baselist+=[_winreg.EnumValue(key,i)[1]]
-				i+=1
-		except:
-			pass
-	baselist = baselist \
-			   + glob.glob("/dev/ttyUSB*") \
-			   + glob.glob("/dev/ttyACM*") \
-			   + glob.glob("/dev/tty.usb*") \
-			   + glob.glob("/dev/cu.*") \
-			   + glob.glob("/dev/cuaU*") \
-			   + glob.glob("/dev/rfcomm*")
-
-	additionalPorts = settings().get(["serial", "additionalPorts"])
-	for additional in additionalPorts:
-		baselist += glob.glob(additional)
-
-	prev = settings().get(["serial", "port"])
-	if prev in baselist:
-		baselist.remove(prev)
-		baselist.insert(0, prev)
-	if settings().getBoolean(["devel", "virtualPrinter", "enabled"]):
-		baselist.append("VIRTUAL")
-	return baselist
-
-def baudrateList():
-	ret = [250000, 230400, 115200, 57600, 38400, 19200, 9600]
-	prev = settings().getInt(["serial", "baudrate"])
-	if prev in ret:
-		ret.remove(prev)
-		ret.insert(0, prev)
-	return ret
-
-gcodeToEvent = {
-	# pause for user input
-	"M226": Events.WAITING,
-	"M0": Events.WAITING,
-	"M1": Events.WAITING,
-	# dwell command
-	"G4": Events.DWELL,
-
-	# part cooler
-	"M245": Events.COOLING,
-
-	# part conveyor
-	"M240": Events.CONVEYOR,
-
-	# part ejector
-	"M40": Events.EJECT,
-
-	# user alert
-	"M300": Events.ALERT,
-
-	# home print head
-	"G28": Events.HOME,
-
-	# emergency stop
-	"M112": Events.E_STOP,
-
-	# motors on/off
-	"M80": Events.POWER_ON,
-	"M81": Events.POWER_OFF,
-}
+from octoprint.settings import settings, default_settings
+from octoprint.util import (get_exception_string, sanitize_ascii, filter_non_ascii,
+							CountedEvent, RepeatedTimer, comm_helpers)
 
 class MachineCom(object):
 	STATE_NONE = 0
@@ -134,7 +61,7 @@ class MachineCom(object):
 		self._printerProfileManager = printerProfileManager
 		self._state = self.STATE_NONE
 		self._serial = None
-		self._baudrateDetectList = baudrateList()
+		self._baudrateDetectList = comm_helpers.baudrateList()
 		self._baudrateDetectRetry = 0
 		self._temp = {}
 		self._bedTemp = None
@@ -164,7 +91,7 @@ class MachineCom(object):
 		self._resendSwallowNextOk = False
 
 		self._clear_to_send = CountedEvent(max=10, name="comm.clear_to_send")
-		self._send_queue = TypedQueue()
+		self._send_queue = comm_helpers.TypedQueue()
 		self._temperature_timer = None
 		self._sd_status_timer = None
 
@@ -320,10 +247,10 @@ class MachineCom(object):
 		return self.isSdFileSelected() and self.isPrinting()
 
 	def isSdFileSelected(self):
-		return self._currentFile is not None and isinstance(self._currentFile, PrintingSdFileInformation)
+		return self._currentFile is not None and isinstance(self._currentFile, comm_helpers.PrintingSdFileInformation)
 
 	def isStreaming(self):
-		return self._currentFile is not None and isinstance(self._currentFile, StreamingGcodeFileInformation)
+		return self._currentFile is not None and isinstance(self._currentFile, comm_helpers.StreamingGcodeFileInformation)
 
 	def isPaused(self):
 		return self._state == self.STATE_PAUSED
@@ -428,7 +355,7 @@ class MachineCom(object):
 	def sendCommand(self, cmd, cmd_type=None, processed=False):
 		cmd = cmd.encode('ascii', 'replace')
 		if not processed:
-			cmd = process_gcode_line(cmd)
+			cmd = comm_helpers.process_gcode_line(cmd)
 			if not cmd:
 				return
 
@@ -452,7 +379,7 @@ class MachineCom(object):
 			scriptLines = filter(
 				lambda x: x is not None and x.strip() != "",
 				map(
-					lambda x: process_gcode_line(x, offsets=self._tempOffsets, current_tool=self._currentTool),
+					lambda x: comm_helpers.process_gcode_line(x, offsets=self._tempOffsets, current_tool=self._currentTool),
 					template.split("\n")
 				)
 			)
@@ -527,7 +454,7 @@ class MachineCom(object):
 
 				self.sendCommand("M24")
 
-				self._sd_status_timer = RepeatedTimer(lambda: get_interval("sdStatus", default_value=1.0), self._poll_sd_status, run_first=True)
+				self._sd_status_timer = RepeatedTimer(lambda: comm_helpers.get_interval("sdStatus", default_value=1.0), self._poll_sd_status, run_first=True)
 				self._sd_status_timer.start()
 			else:
 				line = self._getNext()
@@ -547,7 +474,7 @@ class MachineCom(object):
 			logging.info("Printer is not operation or busy")
 			return
 
-		self._currentFile = StreamingGcodeFileInformation(filename, localFilename, remoteFilename)
+		self._currentFile = comm_helpers.StreamingGcodeFileInformation(filename, localFilename, remoteFilename)
 		self._currentFile.start()
 
 		self.sendCommand("M28 %s" % remoteFilename)
@@ -565,7 +492,7 @@ class MachineCom(object):
 			self._sdFileToSelect = filename
 			self.sendCommand("M23 %s" % filename)
 		else:
-			self._currentFile = PrintingGcodeFileInformation(filename, offsets_callback=self.getOffsets, current_tool_callback=self.getCurrentTool)
+			self._currentFile = comm_helpers.PrintingGcodeFileInformation(filename, offsets_callback=self.getOffsets, current_tool_callback=self.getCurrentTool)
 			eventManager().fire(Events.FILE_SELECTED, {
 				"file": self._currentFile.getFilename(),
 				"filename": os.path.basename(self._currentFile.getFilename()),
@@ -670,7 +597,7 @@ class MachineCom(object):
 
 	def deleteSdFile(self, filename):
 		if not self.isOperational() or (self.isBusy() and
-				isinstance(self._currentFile, PrintingSdFileInformation) and
+				isinstance(self._currentFile, comm_helpers.PrintingSdFileInformation) and
 				self._currentFile.getFilename() == filename):
 			# do not delete a file from sd we are currently printing from
 			return
@@ -781,9 +708,9 @@ class MachineCom(object):
 	##~~ Serial monitor processing received messages
 
 	def _monitor(self):
-		feedback_controls, feedback_matcher = convert_feedback_controls(settings().get(["controls"]))
+		feedback_controls, feedback_matcher = comm_helpers.convert_feedback_controls(settings().get(["controls"]))
 		feedback_errors = []
-		pause_triggers = convert_pause_triggers(settings().get(["printerParameters", "pauseTriggers"]))
+		pause_triggers = comm_helpers.convert_pause_triggers(settings().get(["printerParameters", "pauseTriggers"]))
 
 		disable_external_heatup_detection = not settings().getBoolean(["feature", "externalHeatupDetection"])
 
@@ -803,7 +730,7 @@ class MachineCom(object):
 			self._changeState(self.STATE_CONNECTING)
 
 		#Start monitoring the serial port.
-		self._timeout = get_new_timeout("communication")
+		self._timeout = comm_helpers.get_new_timeout("communication")
 
 		startSeen = False
 		supportRepetierTargetTemp = settings().getBoolean(["feature", "repetierTargetTemp"])
@@ -823,7 +750,7 @@ class MachineCom(object):
 				if line is None:
 					break
 				if line.strip() is not "":
-					self._timeout = get_new_timeout("communication")
+					self._timeout = comm_helpers.get_new_timeout("communication")
 
 				##~~ debugging output handling
 				if line.startswith("//"):
@@ -963,7 +890,7 @@ class MachineCom(object):
 						self._sdFileToSelect = None
 					else:
 						name = match.group(1)
-					self._currentFile = PrintingSdFileInformation(name, int(match.group(2)))
+					self._currentFile = comm_helpers.PrintingSdFileInformation(name, int(match.group(2)))
 				elif 'File selected' in line:
 					if self._ignore_select:
 						self._ignore_select = False
@@ -1046,7 +973,7 @@ class MachineCom(object):
 									self._serial.timeout = connection_timeout
 								self._log("Trying baudrate: %d" % (baudrate))
 								self._baudrateDetectRetry = 5
-								self._timeout = get_new_timeout("communication")
+								self._timeout = comm_helpers.get_new_timeout("communication")
 								self._serial.write('\n')
 								self._sendCommand("M110")
 								self._clear_to_send.set()
@@ -1183,7 +1110,7 @@ class MachineCom(object):
 
 	def _onConnected(self):
 		self._serial.timeout = settings().getFloat(["serial", "timeout", "communication"])
-		self._temperature_timer = RepeatedTimer(lambda: get_interval("temperature", default_value=4.0), self._poll_temperature, run_first=True)
+		self._temperature_timer = RepeatedTimer(lambda: comm_helpers.get_interval("temperature", default_value=4.0), self._poll_temperature, run_first=True)
 		self._temperature_timer.start()
 
 		self._changeState(self.STATE_OPERATIONAL)
@@ -1215,8 +1142,8 @@ class MachineCom(object):
 
 	def _detectPort(self, close):
 		programmer = stk500v2.Stk500v2()
-		self._log("Serial port list: %s" % (str(serialList())))
-		for p in serialList():
+		self._log("Serial port list: %s" % (str(comm_helpers.serialList())))
+		for p in comm_helpers.serialList():
 			serial_obj = None
 
 			try:
@@ -1254,7 +1181,7 @@ class MachineCom(object):
 			# connect to regular serial port
 			self._log("Connecting to: %s" % port)
 			if baudrate == 0:
-				baudrates = baudrateList()
+				baudrates = comm_helpers.baudrateList()
 				serial_obj = serial.Serial(str(port), 115200 if 115200 in baudrates else baudrates[0], timeout=read_timeout, writeTimeout=10000, parity=serial.PARITY_ODD)
 			else:
 				serial_obj = serial.Serial(str(port), baudrate, timeout=read_timeout, writeTimeout=10000, parity=serial.PARITY_ODD)
@@ -1447,9 +1374,9 @@ class MachineCom(object):
 					# command is no more, return
 					return
 
-				if gcode and gcode in gcodeToEvent:
+				if gcode and gcode in comm_helpers.gcodeToEvent:
 					# if this is a gcode bound to an event, trigger that now
-					eventManager().fire(gcodeToEvent[gcode])
+					eventManager().fire(comm_helpers.gcodeToEvent[gcode])
 
 			# actually enqueue the command for sending
 			self._enqueue_for_sending(cmd, command_type=cmd_type)
@@ -1491,7 +1418,7 @@ class MachineCom(object):
 
 		try:
 			self._send_queue.put((command, linenumber, command_type))
-		except TypeAlreadyInQueue as e:
+		except comm_helpers.TypeAlreadyInQueue as e:
 			self._logger.debug("Type already in queue: " + e.type)
 
 	def _send_loop(self):
@@ -1750,351 +1677,11 @@ class MachineCom(object):
 		elif s_idx != -1:
 			# dwell time is specified in seconds
 			_timeout = float(cmd[s_idx+1:])
-		self._timeout = get_new_timeout("communication") + _timeout
+		self._timeout = comm_helpers.get_new_timeout("communication") + _timeout
 
 	##~~ command phase handlers
 
 	def _command_phase_sending(self, cmd, cmd_type=None, gcode=None):
 		if gcode is not None and gcode in self._long_running_commands:
 			self._long_running_command = True
-
-
-### Printing file information classes ##################################################################################
-
-class PrintingFileInformation(object):
-	"""
-	Encapsulates information regarding the current file being printed: file name, current position, total size and
-	time the print started.
-	Allows to reset the current file position to 0 and to calculate the current progress as a floating point
-	value between 0 and 1.
-	"""
-
-	def __init__(self, filename):
-		self._logger = logging.getLogger(__name__)
-		self._filename = filename
-		self._pos = 0
-		self._size = None
-		self._start_time = None
-
-	def getStartTime(self):
-		return self._start_time
-
-	def getFilename(self):
-		return self._filename
-
-	def getFilesize(self):
-		return self._size
-
-	def getFilepos(self):
-		return self._pos
-
-	def getFileLocation(self):
-		return FileDestinations.LOCAL
-
-	def getProgress(self):
-		"""
-		The current progress of the file, calculated as relation between file position and absolute size. Returns -1
-		if file size is None or < 1.
-		"""
-		if self._size is None or not self._size > 0:
-			return -1
-		return float(self._pos) / float(self._size)
-
-	def reset(self):
-		"""
-		Resets the current file position to 0.
-		"""
-		self._pos = 0
-
-	def start(self):
-		"""
-		Marks the print job as started and remembers the start time.
-		"""
-		self._start_time = time.time()
-
-	def close(self):
-		"""
-		Closes the print job.
-		"""
-		pass
-
-class PrintingSdFileInformation(PrintingFileInformation):
-	"""
-	Encapsulates information regarding an ongoing print from SD.
-	"""
-
-	def __init__(self, filename, size):
-		PrintingFileInformation.__init__(self, filename)
-		self._size = size
-
-	def setFilepos(self, pos):
-		"""
-		Sets the current file position.
-		"""
-		self._pos = pos
-
-	def getFileLocation(self):
-		return FileDestinations.SDCARD
-
-class PrintingGcodeFileInformation(PrintingFileInformation):
-	"""
-	Encapsulates information regarding an ongoing direct print. Takes care of the needed file handle and ensures
-	that the file is closed in case of an error.
-	"""
-
-	def __init__(self, filename, offsets_callback=None, current_tool_callback=None):
-		PrintingFileInformation.__init__(self, filename)
-
-		self._handle = None
-
-		self._first_line = None
-
-		self._offsets_callback = offsets_callback
-		self._current_tool_callback = current_tool_callback
-
-		if not os.path.exists(self._filename) or not os.path.isfile(self._filename):
-			raise IOError("File %s does not exist" % self._filename)
-		self._size = os.stat(self._filename).st_size
-		self._pos = 0
-
-	def start(self):
-		"""
-		Opens the file for reading and determines the file size.
-		"""
-		PrintingFileInformation.start(self)
-		self._handle = open(self._filename, "r")
-
-	def close(self):
-		"""
-		Closes the file if it's still open.
-		"""
-		PrintingFileInformation.close(self)
-		if self._handle is not None:
-			try:
-				self._handle.close()
-			except:
-				pass
-		self._handle = None
-
-	def getNext(self):
-		"""
-		Retrieves the next line for printing.
-		"""
-		if self._handle is None:
-			raise ValueError("File %s is not open for reading" % self._filename)
-
-		try:
-			offsets = self._offsets_callback() if self._offsets_callback is not None else None
-			current_tool = self._current_tool_callback() if self._current_tool_callback is not None else None
-
-			processed = None
-			while processed is None:
-				if self._handle is None:
-					# file got closed just now
-					return None
-				line = self._handle.readline()
-				if not line:
-					self.close()
-				processed = process_gcode_line(line, offsets=offsets, current_tool=current_tool)
-			self._pos = self._handle.tell()
-
-			return processed
-		except Exception as e:
-			self.close()
-			self._logger.exception("Exception while processing line")
-			raise e
-
-class StreamingGcodeFileInformation(PrintingGcodeFileInformation):
-	def __init__(self, path, localFilename, remoteFilename):
-		PrintingGcodeFileInformation.__init__(self, path)
-		self._localFilename = localFilename
-		self._remoteFilename = remoteFilename
-
-	def start(self):
-		PrintingGcodeFileInformation.start(self)
-		self._start_time = time.time()
-
-	def getLocalFilename(self):
-		return self._localFilename
-
-	def getRemoteFilename(self):
-		return self._remoteFilename
-
-
-class TypedQueue(queue.Queue):
-
-	def __init__(self, maxsize=0):
-		queue.Queue.__init__(self, maxsize=maxsize)
-		self._lookup = []
-
-	def _put(self, item):
-		if isinstance(item, tuple) and len(item) == 3:
-			cmd, line, cmd_type = item
-			if cmd_type is not None:
-				if cmd_type in self._lookup:
-					raise TypeAlreadyInQueue(cmd_type, "Type {cmd_type} is already in queue".format(**locals()))
-				else:
-					self._lookup.append(cmd_type)
-
-		queue.Queue._put(self, item)
-
-	def _get(self):
-		item = queue.Queue._get(self)
-
-		if isinstance(item, tuple) and len(item) == 3:
-			cmd, line, cmd_type = item
-			if cmd_type is not None and cmd_type in self._lookup:
-				self._lookup.remove(cmd_type)
-
-		return item
-
-
-class TypeAlreadyInQueue(Exception):
-	def __init__(self, t, *args, **kwargs):
-		Exception.__init__(self, *args, **kwargs)
-		self.type = t
-
-
-def get_new_timeout(type):
-	now = time.time()
-	return now + get_interval(type)
-
-
-def get_interval(type, default_value=0.0):
-	if type not in default_settings["serial"]["timeout"]:
-		return default_value
-	else:
-		value = settings().getFloat(["serial", "timeout", type])
-		if not value:
-			return default_value
-		else:
-			return value
-
-_temp_command_regex = re.compile("^M(?P<command>104|109|140|190)(\s+T(?P<tool>\d+)|\s+S(?P<temperature>[-+]?\d*\.?\d*))+")
-
-def apply_temperature_offsets(line, offsets, current_tool=None):
-	if offsets is None:
-		return line
-
-	match = _temp_command_regex.match(line)
-	if match is None:
-		return line
-
-	groups = match.groupdict()
-	if not "temperature" in groups or groups["temperature"] is None:
-		return line
-
-	offset = 0
-	if current_tool is not None and (groups["command"] == "104" or groups["command"] == "109"):
-		# extruder temperature, determine which one and retrieve corresponding offset
-		tool_num = current_tool
-		if "tool" in groups and groups["tool"] is not None:
-			tool_num = int(groups["tool"])
-
-		tool_key = "tool%d" % tool_num
-		offset = offsets[tool_key] if tool_key in offsets and offsets[tool_key] else 0
-
-	elif groups["command"] == "140" or groups["command"] == "190":
-		# bed temperature
-		offset = offsets["bed"] if "bed" in offsets else 0
-
-	if offset == 0:
-		return line
-
-	temperature = float(groups["temperature"])
-	if temperature == 0:
-		return line
-
-	return line[:match.start("temperature")] + "%f" % (temperature + offset) + line[match.end("temperature"):]
-
-def strip_comment(line):
-	if not ";" in line:
-		# shortcut
-		return line
-
-	escaped = False
-	result = []
-	for c in line:
-		if c == ";" and not escaped:
-			break
-		result += c
-		escaped = (c == "\\") and not escaped
-	return "".join(result)
-
-def process_gcode_line(line, offsets=None, current_tool=None):
-	line = strip_comment(line).strip()
-	if not len(line):
-		return None
-
-	if offsets is not None:
-		line = apply_temperature_offsets(line, offsets, current_tool=current_tool)
-
-	return line
-
-def convert_pause_triggers(configured_triggers):
-	triggers = {
-		"enable": [],
-		"disable": [],
-		"toggle": []
-	}
-	for trigger in configured_triggers:
-		if not "regex" in trigger or not "type" in trigger:
-			continue
-
-		try:
-			regex = trigger["regex"]
-			t = trigger["type"]
-			if t in triggers:
-				# make sure regex is valid
-				re.compile(regex)
-				# add to type list
-				triggers[t].append(regex)
-		except:
-			# invalid regex or something like this, we'll just skip this entry
-			pass
-
-	result = dict()
-	for t in triggers.keys():
-		if len(triggers[t]) > 0:
-			result[t] = re.compile("|".join(map(lambda pattern: "({pattern})".format(pattern=pattern), triggers[t])))
-	return result
-
-
-def convert_feedback_controls(configured_controls):
-	def preprocess_feedback_control(control, result):
-		if "key" in control and "regex" in control and "template" in control:
-			# key is always the md5sum of the regex
-			key = control["key"]
-
-			if result[key]["pattern"] is None or result[key]["matcher"] is None:
-				# regex has not been registered
-				try:
-					result[key]["matcher"] = re.compile(control["regex"])
-					result[key]["pattern"] = control["regex"]
-				except Exception as exc:
-					logging.getLogger(__name__).warn("Invalid regex {regex} for custom control: {exc}".format(regex=control["regex"], exc=str(exc)))
-
-			result[key]["templates"][control["template_key"]] = control["template"]
-
-		elif "children" in control:
-			for c in control["children"]:
-				preprocess_feedback_control(c, result)
-
-	def prepare_result_entry():
-		return dict(pattern=None, matcher=None, templates=dict())
-
-	from collections import defaultdict
-	feedback_controls = defaultdict(prepare_result_entry)
-
-	for control in configured_controls:
-		preprocess_feedback_control(control, feedback_controls)
-
-	feedback_pattern = []
-	for match_key, entry in feedback_controls.items():
-		if entry["matcher"] is None or entry["pattern"] is None:
-			continue
-		feedback_pattern.append("(?P<group{key}>{pattern})".format(key=match_key, pattern=entry["pattern"]))
-	feedback_matcher = re.compile("|".join(feedback_pattern))
-
-	return feedback_controls, feedback_matcher
