+++ src/octoprint/printer/__init__.py
@@ -20,8 +20,8 @@ __copyright__ = "Copyright (C) 2014 The OctoPrint Project - Released under terms
 
 import re
 
-import octoprint.util.comm as comm
 from octoprint.settings import settings
+from octoprint.util import comm_helpers
 
 def get_connection_options():
 	"""
@@ -39,8 +39,8 @@ def get_connection_options():
 	    (dict): A dictionary holding the connection options in the structure specified above
 	"""
 	return {
-		"ports": comm.serialList(),
-		"baudrates": comm.baudrateList(),
+		"ports": comm_helpers.serialList(),
+		"baudrates": comm_helpers.baudrateList(),
 		"portPreference": settings().get(["serial", "port"]),
 		"baudratePreference": settings().getInt(["serial", "baudrate"]),
 		"autoconnect": settings().getBoolean(["serial", "autoconnect"])
